# [static runtime] Silence deprecation warnings with resize_